+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Minor Changes
 -------------
 
 - sir_config - Add support for append `eof` to the end of backup configuration.
+- sir_config - Add support for `commit try` command.
 
 v1.1.0
 ======